Refactor munge fn from <TrendContainer /> to summary util
