+++ src/components/TrendContainer.js
@@ -7,25 +7,8 @@ import Loading from './Loading'
 import NoData from './NoData'
 import TrendChart from './TrendChart'
 import TrendSourceText from './TrendSourceText'
+import mungeSummaryData from '../util/summary'
 
-
-const mungeSummaryData = (crime, summaries, place) => {
-  if (!summaries || !summaries[place]) return false
-
-  const keys = Object.keys(summaries)
-  return summaries[place].map(year => {
-    const data = { date: year.year }
-    keys.forEach(key => {
-      const source = key !== place ? summaries[key].find(d => d.year === data.date) : year
-      data[key] = {
-        pop: source.population,
-        count: source[crime],
-        rate: (source[crime] / source.population) * 100000,
-      }
-    })
-    return data
-  })
-}
 
 const TrendContainer = ({
   crime,
